feat : 몽고디비 초기 설정
